Module Make.Script_typed_ir

type var_annot =

| Var_annot of string

type type_annot =

| Type_annot of string

type field_annot =

| Field_annot of string

type never =

type address = Alpha_context.Contract.t * string

type ('a, 'b) pair = 'a * 'b

type ('a, 'b) union =

| L of 'a

| R of 'b

type operation = Alpha_context.packed_internal_operation * Alpha_context.Lazy_storage.diffs option

type 'a ticket = {

ticketer : address;

contents : 'a;

amount : Alpha_context.Script_int.n Alpha_context.Script_int.num;

}

type empty_cell =

| EmptyCell

type end_of_stack = empty_cell * empty_cell

type _ comparable_ty =

| Unit_key : type_annot option -> unit comparable_ty

| Never_key : type_annot option -> never comparable_ty

| Int_key : type_annot option -> Alpha_context.Script_int.z Alpha_context.Script_int.num comparable_ty

| Nat_key : type_annot option -> Alpha_context.Script_int.n Alpha_context.Script_int.num comparable_ty

| Signature_key : type_annot option -> Alpha_context.signature comparable_ty

| String_key : type_annot option -> string comparable_ty

| Bytes_key : type_annot option -> Tezos_protocol_environment.Bytes.t comparable_ty

| Mutez_key : type_annot option -> Alpha_context.Tez.t comparable_ty

| Bool_key : type_annot option -> bool comparable_ty

| Key_hash_key : type_annot option -> Alpha_context.public_key_hash comparable_ty

| Key_key : type_annot option -> Alpha_context.public_key comparable_ty

| Timestamp_key : type_annot option -> Alpha_context.Script_timestamp.t comparable_ty

| Chain_id_key : type_annot option -> Tezos_protocol_environment.Chain_id.t comparable_ty

| Address_key : type_annot option -> address comparable_ty

| Pair_key : ('a comparable_ty * field_annot option) * ('b comparable_ty * field_annot option) * type_annot option -> ('a, 'b) pair comparable_ty

| Union_key : ('a comparable_ty * field_annot option) * ('b comparable_ty * field_annot option) * type_annot option -> ('a, 'b) union comparable_ty

| Option_key : 'v comparable_ty * type_annot option -> 'v option comparable_ty

module type Boxed_set = sig ... end

type 'elt set = (module Boxed_set with type elt = 'elt)

module type Boxed_map = sig ... end

type ('key, 'value) map = (module Boxed_map with type key = 'key and type value = 'value)

module Big_map_overlay : sig ... end

type ('key, 'value) big_map_overlay = {

map : ('key * 'value option) Big_map_overlay.t;

size : int;

}

type 'elt boxed_list = {

elements : 'elt list;

length : int;

}

type ('arg, 'storage) script = {

code : (('arg, 'storage) pair, (operation boxed_list, 'storage) pair) lambda;

arg_type : 'arg ty;

storage : 'storage;

storage_type : 'storage ty;

root_name : field_annot option;

}

and ('before_top, 'before, 'result_top, 'result) kinstr =

| IDrop : ('a, 'b * 's) kinfo * ('b, 's, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IDup : ('a, 's) kinfo * ('a, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISwap : ('a, 'b * 's) kinfo * ('b, 'a * 's, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IConst : ('a, 's) kinfo * 'ty * ('ty, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ICons_pair : ('a, 'b * 's) kinfo * ('a * 'b, 's, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| ICar : ('a * 'b, 's) kinfo * ('a, 's, 'r, 'f) kinstr -> ('a * 'b, 's, 'r, 'f) kinstr

| ICdr : ('a * 'b, 's) kinfo * ('b, 's, 'r, 'f) kinstr -> ('a * 'b, 's, 'r, 'f) kinstr

| IUnpair : ('a * 'b, 's) kinfo * ('a, 'b * 's, 'r, 'f) kinstr -> ('a * 'b, 's, 'r, 'f) kinstr

| ICons_some : ('v, 's) kinfo * ('v option, 's, 'r, 'f) kinstr -> ('v, 's, 'r, 'f) kinstr

| ICons_none : ('a, 's) kinfo * 'b ty * ('b option, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IIf_none : {

kinfo : ('a option, 'b * 's) kinfo;

branch_if_none : ('b, 's, 'r, 'f) kinstr;

branch_if_some : ('a, 'b * 's, 'r, 'f) kinstr;

} -> ('a option, 'b * 's, 'r, 'f) kinstr

| ICons_left : ('a, 's) kinfo * (('a, 'b) union, 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ICons_right : ('b, 's) kinfo * (('a, 'b) union, 's, 'r, 'f) kinstr -> ('b, 's, 'r, 'f) kinstr

| IIf_left : {

kinfo : (('a, 'b) union, 's) kinfo;

branch_if_left : ('a, 's, 'r, 'f) kinstr;

branch_if_right : ('b, 's, 'r, 'f) kinstr;

} -> (('a, 'b) union, 's, 'r, 'f) kinstr

| ICons_list : ('a, 'a boxed_list * 's) kinfo * ('a boxed_list, 's, 'r, 'f) kinstr -> ('a, 'a boxed_list * 's, 'r, 'f) kinstr

| INil : ('a, 's) kinfo * ('b boxed_list, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IIf_cons : {

kinfo : ('a boxed_list, 'b * 's) kinfo;

branch_if_cons : ('a, 'a boxed_list * ('b * 's), 'r, 'f) kinstr;

branch_if_nil : ('b, 's, 'r, 'f) kinstr;

} -> ('a boxed_list, 'b * 's, 'r, 'f) kinstr

| IList_map : ('a boxed_list, 'c * 's) kinfo * ('a, 'c * 's, 'b, 'c * 's) kinstr * ('b boxed_list, 'c * 's, 'r, 'f) kinstr -> ('a boxed_list, 'c * 's, 'r, 'f) kinstr

| IList_iter : ('a boxed_list, 'b * 's) kinfo * ('a, 'b * 's, 'b, 's) kinstr * ('b, 's, 'r, 'f) kinstr -> ('a boxed_list, 'b * 's, 'r, 'f) kinstr

| IList_size : ('a boxed_list, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a boxed_list, 's, 'r, 'f) kinstr

| IEmpty_set : ('a, 's) kinfo * 'b comparable_ty * ('b set, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISet_iter : ('a set, 'b * 's) kinfo * ('a, 'b * 's, 'b, 's) kinstr * ('b, 's, 'r, 'f) kinstr -> ('a set, 'b * 's, 'r, 'f) kinstr

| ISet_mem : ('a, 'a set * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> ('a, 'a set * 's, 'r, 'f) kinstr

| ISet_update : ('a, bool * ('a set * 's)) kinfo * ('a set, 's, 'r, 'f) kinstr -> ('a, bool * ('a set * 's), 'r, 'f) kinstr

| ISet_size : ('a set, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a set, 's, 'r, 'f) kinstr

| IEmpty_map : ('a, 's) kinfo * 'b comparable_ty * 'c ty * (('b, 'c) map, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IMap_map : (('a, 'b) map, 'd * 's) kinfo * ('a * 'b, 'd * 's, 'c, 'd * 's) kinstr * (('a, 'c) map, 'd * 's, 'r, 'f) kinstr -> (('a, 'b) map, 'd * 's, 'r, 'f) kinstr

| IMap_iter : (('a, 'b) map, 'c * 's) kinfo * ('a * 'b, 'c * 's, 'c, 's) kinstr * ('c, 's, 'r, 'f) kinstr -> (('a, 'b) map, 'c * 's, 'r, 'f) kinstr

| IMap_mem : ('a, ('a, 'b) map * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) map * 's, 'r, 'f) kinstr

| IMap_get : ('a, ('a, 'b) map * 's) kinfo * ('b option, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) map * 's, 'r, 'f) kinstr

| IMap_update : ('a, 'b option * (('a, 'b) map * 's)) kinfo * (('a, 'b) map, 's, 'r, 'f) kinstr -> ('a, 'b option * (('a, 'b) map * 's), 'r, 'f) kinstr

| IMap_get_and_update : ('a, 'b option * (('a, 'b) map * 's)) kinfo * ('b option, ('a, 'b) map * 's, 'r, 'f) kinstr -> ('a, 'b option * (('a, 'b) map * 's), 'r, 'f) kinstr

| IMap_size : (('a, 'b) map, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (('a, 'b) map, 's, 'r, 'f) kinstr

| IEmpty_big_map : ('a, 's) kinfo * 'b comparable_ty * 'c ty * (('b, 'c) big_map, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IBig_map_mem : ('a, ('a, 'b) big_map * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) big_map * 's, 'r, 'f) kinstr

| IBig_map_get : ('a, ('a, 'b) big_map * 's) kinfo * ('b option, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) big_map * 's, 'r, 'f) kinstr

| IBig_map_update : ('a, 'b option * (('a, 'b) big_map * 's)) kinfo * (('a, 'b) big_map, 's, 'r, 'f) kinstr -> ('a, 'b option * (('a, 'b) big_map * 's), 'r, 'f) kinstr

| IBig_map_get_and_update : ('a, 'b option * (('a, 'b) big_map * 's)) kinfo * ('b option, ('a, 'b) big_map * 's, 'r, 'f) kinstr -> ('a, 'b option * (('a, 'b) big_map * 's), 'r, 'f) kinstr

| IConcat_string : (string boxed_list, 's) kinfo * (string, 's, 'r, 'f) kinstr -> (string boxed_list, 's, 'r, 'f) kinstr

| IConcat_string_pair : (string, string * 's) kinfo * (string, 's, 'r, 'f) kinstr -> (string, string * 's, 'r, 'f) kinstr

| ISlice_string : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * (string * 's)) kinfo * (string option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * (string * 's), 'r, 'f) kinstr

| IString_size : (string, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (string, 's, 'r, 'f) kinstr

| IConcat_bytes : (bytes boxed_list, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes boxed_list, 's, 'r, 'f) kinstr

| IConcat_bytes_pair : (bytes, bytes * 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, bytes * 's, 'r, 'f) kinstr

| ISlice_bytes : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * (bytes * 's)) kinfo * (bytes option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * (bytes * 's), 'r, 'f) kinstr

| IBytes_size : (bytes, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| IAdd_seconds_to_timestamp : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_timestamp.t * 's) kinfo * (Alpha_context.Script_timestamp.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_timestamp.t * 's, 'r, 'f) kinstr

| IAdd_timestamp_to_seconds : (Alpha_context.Script_timestamp.t, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_timestamp.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_timestamp.t, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| ISub_timestamp_seconds : (Alpha_context.Script_timestamp.t, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_timestamp.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_timestamp.t, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IDiff_timestamps : (Alpha_context.Script_timestamp.t, Alpha_context.Script_timestamp.t * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_timestamp.t, Alpha_context.Script_timestamp.t * 's, 'r, 'f) kinstr

| IAdd_tez : (Alpha_context.Tez.t, Alpha_context.Tez.t * 's) kinfo * (Alpha_context.Tez.t, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| ISub_tez : (Alpha_context.Tez.t, Alpha_context.Tez.t * 's) kinfo * (Alpha_context.Tez.t, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| IMul_teznat : (Alpha_context.Tez.t, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Tez.t, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_nattez : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Tez.t * 's) kinfo * (Alpha_context.Tez.t, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| IEdiv_teznat : (Alpha_context.Tez.t, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * ((Alpha_context.Tez.t, Alpha_context.Tez.t) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IEdiv_tez : (Alpha_context.Tez.t, Alpha_context.Tez.t * 's) kinfo * ((Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Tez.t) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Tez.t, Alpha_context.Tez.t * 's, 'r, 'f) kinstr

| IOr : (bool, bool * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (bool, bool * 's, 'r, 'f) kinstr

| IAnd : (bool, bool * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (bool, bool * 's, 'r, 'f) kinstr

| IXor : (bool, bool * 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (bool, bool * 's, 'r, 'f) kinstr

| INot : (bool, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (bool, 's, 'r, 'f) kinstr

| IIs_nat : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| INeg_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| INeg_int : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IAbs_int : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IInt_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IAdd_intint : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IAdd_intnat : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IAdd_natint : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IAdd_natnat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| ISub_int : ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, 'b Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_intint : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_intnat : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_natint : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_natnat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IEdiv_intint : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * ((Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IEdiv_intnat : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * ((Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IEdiv_natint : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's) kinfo * ((Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.z Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IEdiv_natnat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * ((Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| ILsl_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| ILsr_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IOr_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IAnd_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IAnd_int_nat : (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IXor_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| INot_nat : (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| INot_int : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IIf : {

kinfo : (bool, 'a * 's) kinfo;

branch_if_true : ('a, 's, 'r, 'f) kinstr;

branch_if_false : ('a, 's, 'r, 'f) kinstr;

} -> (bool, 'a * 's, 'r, 'f) kinstr

| ILoop : (bool, 'a * 's) kinfo * ('a, 's, bool, 'a * 's) kinstr * ('a, 's, 'r, 'f) kinstr -> (bool, 'a * 's, 'r, 'f) kinstr

| ILoop_left : (('a, 'b) union, 's) kinfo * ('a, 's, ('a, 'b) union, 's) kinstr * ('b, 's, 'r, 'f) kinstr -> (('a, 'b) union, 's, 'r, 'f) kinstr

| IDip : ('a, 'b * 's) kinfo * ('b, 's, 'c, 't) kinstr * ('a, 'c * 't, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IExec : ('a, ('a, 'b) lambda * 's) kinfo * ('b, 's, 'r, 'f) kinstr -> ('a, ('a, 'b) lambda * 's, 'r, 'f) kinstr

| IApply : ('a, ('a * 'b, 'c) lambda * 's) kinfo * 'a ty * (('b, 'c) lambda, 's, 'r, 'f) kinstr -> ('a, ('a * 'b, 'c) lambda * 's, 'r, 'f) kinstr

| ILambda : ('a, 's) kinfo * ('b, 'c) lambda * (('b, 'c) lambda, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IFailwith : ('a, 's) kinfo * Alpha_context.Script.location * 'a ty * ('b, 't, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ICompare : ('a, 'a * 's) kinfo * 'a comparable_ty * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> ('a, 'a * 's, 'r, 'f) kinstr

| IEq : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| INeq : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| ILt : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IGt : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| ILe : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IGe : (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr

| IAddress : ('a typed_contract, 's) kinfo * (address, 's, 'r, 'f) kinstr -> ('a typed_contract, 's, 'r, 'f) kinstr

| IContract : (address, 's) kinfo * 'a ty * string * ('a typed_contract option, 's, 'r, 'f) kinstr -> (address, 's, 'r, 'f) kinstr

| ITransfer_tokens : ('a, Alpha_context.Tez.t * ('a typed_contract * 's)) kinfo * (operation, 's, 'r, 'f) kinstr -> ('a, Alpha_context.Tez.t * ('a typed_contract * 's), 'r, 'f) kinstr

| IImplicit_account : (Alpha_context.public_key_hash, 's) kinfo * (unit typed_contract, 's, 'r, 'f) kinstr -> (Alpha_context.public_key_hash, 's, 'r, 'f) kinstr

| ICreate_contract : {

kinfo : (Alpha_context.public_key_hash option, Alpha_context.Tez.t * ('a * 's)) kinfo;

storage_type : 'a ty;

arg_type : 'b ty;

lambda : ('b * 'a, operation boxed_list * 'a) lambda;

root_name : field_annot option;

k : (operation, address * 's, 'r, 'f) kinstr;

} -> (Alpha_context.public_key_hash option, Alpha_context.Tez.t * ('a * 's), 'r, 'f) kinstr

| ISet_delegate : (Alpha_context.public_key_hash option, 's) kinfo * (operation, 's, 'r, 'f) kinstr -> (Alpha_context.public_key_hash option, 's, 'r, 'f) kinstr

| INow : ('a, 's) kinfo * (Alpha_context.Script_timestamp.t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IBalance : ('a, 's) kinfo * (Alpha_context.Tez.t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ILevel : ('a, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ICheck_signature : (Alpha_context.public_key, Alpha_context.signature * (bytes * 's)) kinfo * (bool, 's, 'r, 'f) kinstr -> (Alpha_context.public_key, Alpha_context.signature * (bytes * 's), 'r, 'f) kinstr

| IHash_key : (Alpha_context.public_key, 's) kinfo * (Alpha_context.public_key_hash, 's, 'r, 'f) kinstr -> (Alpha_context.public_key, 's, 'r, 'f) kinstr

| IPack : ('a, 's) kinfo * 'a ty * (bytes, 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IUnpack : (bytes, 's) kinfo * 'a ty * ('a option, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| IBlake2b : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| ISha256 : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| ISha512 : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| ISource : ('a, 's) kinfo * (address, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISender : ('a, 's) kinfo * (address, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISelf : ('a, 's) kinfo * 'b ty * string * ('b typed_contract, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISelf_address : ('a, 's) kinfo * (address, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IAmount : ('a, 's) kinfo * (Alpha_context.Tez.t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ISapling_empty_state : ('a, 's) kinfo * Alpha_context.Sapling.Memo_size.t * (Alpha_context.Sapling.state, 'a * 's, 'b, 'f) kinstr -> ('a, 's, 'b, 'f) kinstr

| ISapling_verify_update : (Alpha_context.Sapling.transaction, Alpha_context.Sapling.state * 's) kinfo * ((Alpha_context.Script_int.z Alpha_context.Script_int.num, Alpha_context.Sapling.state) pair option, 's, 'r, 'f) kinstr -> (Alpha_context.Sapling.transaction, Alpha_context.Sapling.state * 's, 'r, 'f) kinstr

| IDig : ('a, 's) kinfo * int * ('b, 'c * 't, 'c, 't, 'a, 's, 'd, 'u) stack_prefix_preservation_witness * ('b, 'd * 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IDug : ('a, 'b * 's) kinfo * int * ('c, 't, 'a, 'c * 't, 'b, 's, 'd, 'u) stack_prefix_preservation_witness * ('d, 'u, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IDipn : ('a, 's) kinfo * int * ('c, 't, 'd, 'v, 'a, 's, 'b, 'u) stack_prefix_preservation_witness * ('c, 't, 'd, 'v) kinstr * ('b, 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IDropn : ('a, 's) kinfo * int * ('b, 'u, 'b, 'u, 'a, 's, 'a, 's) stack_prefix_preservation_witness * ('b, 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IChainId : ('a, 's) kinfo * (Tezos_protocol_environment.Chain_id.t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| INever : (never, 's) kinfo -> (never, 's, 'r, 'f) kinstr

| IVoting_power : (Alpha_context.public_key_hash, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Alpha_context.public_key_hash, 's, 'r, 'f) kinstr

| ITotal_voting_power : ('a, 's) kinfo * (Alpha_context.Script_int.n Alpha_context.Script_int.num, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IKeccak : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| ISha3 : (bytes, 's) kinfo * (bytes, 's, 'r, 'f) kinstr -> (bytes, 's, 'r, 'f) kinstr

| IAdd_bls12_381_g1 : (Tezos_protocol_environment.Bls12_381.G1.t, Tezos_protocol_environment.Bls12_381.G1.t * 's) kinfo * (Tezos_protocol_environment.Bls12_381.G1.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.G1.t, Tezos_protocol_environment.Bls12_381.G1.t * 's, 'r, 'f) kinstr

| IAdd_bls12_381_g2 : (Tezos_protocol_environment.Bls12_381.G2.t, Tezos_protocol_environment.Bls12_381.G2.t * 's) kinfo * (Tezos_protocol_environment.Bls12_381.G2.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.G2.t, Tezos_protocol_environment.Bls12_381.G2.t * 's, 'r, 'f) kinstr

| IAdd_bls12_381_fr : (Tezos_protocol_environment.Bls12_381.Fr.t, Tezos_protocol_environment.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.Fr.t, Tezos_protocol_environment.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IMul_bls12_381_g1 : (Tezos_protocol_environment.Bls12_381.G1.t, Tezos_protocol_environment.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment.Bls12_381.G1.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.G1.t, Tezos_protocol_environment.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IMul_bls12_381_g2 : (Tezos_protocol_environment.Bls12_381.G2.t, Tezos_protocol_environment.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment.Bls12_381.G2.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.G2.t, Tezos_protocol_environment.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IMul_bls12_381_fr : (Tezos_protocol_environment.Bls12_381.Fr.t, Tezos_protocol_environment.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.Fr.t, Tezos_protocol_environment.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IMul_bls12_381_z_fr : (Tezos_protocol_environment.Bls12_381.Fr.t, 'a Alpha_context.Script_int.num * 's) kinfo * (Tezos_protocol_environment.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.Fr.t, 'a Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IMul_bls12_381_fr_z : ('a Alpha_context.Script_int.num, Tezos_protocol_environment.Bls12_381.Fr.t * 's) kinfo * (Tezos_protocol_environment.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> ('a Alpha_context.Script_int.num, Tezos_protocol_environment.Bls12_381.Fr.t * 's, 'r, 'f) kinstr

| IInt_bls12_381_fr : (Tezos_protocol_environment.Bls12_381.Fr.t, 's) kinfo * (Alpha_context.Script_int.z Alpha_context.Script_int.num, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.Fr.t, 's, 'r, 'f) kinstr

| INeg_bls12_381_g1 : (Tezos_protocol_environment.Bls12_381.G1.t, 's) kinfo * (Tezos_protocol_environment.Bls12_381.G1.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.G1.t, 's, 'r, 'f) kinstr

| INeg_bls12_381_g2 : (Tezos_protocol_environment.Bls12_381.G2.t, 's) kinfo * (Tezos_protocol_environment.Bls12_381.G2.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.G2.t, 's, 'r, 'f) kinstr

| INeg_bls12_381_fr : (Tezos_protocol_environment.Bls12_381.Fr.t, 's) kinfo * (Tezos_protocol_environment.Bls12_381.Fr.t, 's, 'r, 'f) kinstr -> (Tezos_protocol_environment.Bls12_381.Fr.t, 's, 'r, 'f) kinstr

| IPairing_check_bls12_381 : ((Tezos_protocol_environment.Bls12_381.G1.t, Tezos_protocol_environment.Bls12_381.G2.t) pair boxed_list, 's) kinfo * (bool, 's, 'r, 'f) kinstr -> ((Tezos_protocol_environment.Bls12_381.G1.t, Tezos_protocol_environment.Bls12_381.G2.t) pair boxed_list, 's, 'r, 'f) kinstr

| IComb : ('a, 's) kinfo * int * ('a * 's, 'b * 'u) comb_gadt_witness * ('b, 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IUncomb : ('a, 's) kinfo * int * ('a * 's, 'b * 'u) uncomb_gadt_witness * ('b, 'u, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| IComb_get : ('t, 's) kinfo * int * ('t, 'v) comb_get_gadt_witness * ('v, 's, 'r, 'f) kinstr -> ('t, 's, 'r, 'f) kinstr

| IComb_set : ('a, 'b * 's) kinfo * int * ('a, 'b, 'c) comb_set_gadt_witness * ('c, 's, 'r, 'f) kinstr -> ('a, 'b * 's, 'r, 'f) kinstr

| IDup_n : ('a, 's) kinfo * int * ('a * 's, 't) dup_n_gadt_witness * ('t, 'a * 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

| ITicket : ('a, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's) kinfo * ('a ticket, 's, 'r, 'f) kinstr -> ('a, Alpha_context.Script_int.n Alpha_context.Script_int.num * 's, 'r, 'f) kinstr

| IRead_ticket : ('a ticket, 's) kinfo * (address * ('a * Alpha_context.Script_int.n Alpha_context.Script_int.num), 'a ticket * 's, 'r, 'f) kinstr -> ('a ticket, 's, 'r, 'f) kinstr

| ISplit_ticket : ('a ticket, (Alpha_context.Script_int.n Alpha_context.Script_int.num * Alpha_context.Script_int.n Alpha_context.Script_int.num) * 's) kinfo * (('a ticket * 'a ticket) option, 's, 'r, 'f) kinstr -> ('a ticket, (Alpha_context.Script_int.n Alpha_context.Script_int.num * Alpha_context.Script_int.n Alpha_context.Script_int.num) * 's, 'r, 'f) kinstr

| IJoin_tickets : ('a ticket * 'a ticket, 's) kinfo * 'a comparable_ty * ('a ticket option, 's, 'r, 'f) kinstr -> ('a ticket * 'a ticket, 's, 'r, 'f) kinstr

| IHalt : ('a, 's) kinfo -> ('a, 's, 'a, 's) kinstr

| ILog : ('a, 's) kinfo * logging_event * logger * ('a, 's, 'r, 'f) kinstr -> ('a, 's, 'r, 'f) kinstr

and logging_event =

| LogEntry : logging_event

| LogExit : ('b, 'u) kinfo -> logging_event

and ('arg, 'ret) lambda =

| Lam : ('arg, end_of_stack, 'ret, end_of_stack) kdescr * Alpha_context.Script.node -> ('arg, 'ret) lambda

and 'arg typed_contract = 'arg ty * address

and (_, _, _, _) continuation =

| KNil : ('r, 'f, 'r, 'f) continuation

| KCons : ('a, 's, 'b, 't) kinstr * ('b, 't, 'r, 'f) continuation -> ('a, 's, 'r, 'f) continuation

| KReturn : 's * ('a, 's, 'r, 'f) continuation -> ('a, end_of_stack, 'r, 'f) continuation

| KUndip : 'b * ('b, 'a * 's, 'r, 'f) continuation -> ('a, 's, 'r, 'f) continuation

| KLoop_in : ('a, 's, bool, 'a * 's) kinstr * ('a, 's, 'r, 'f) continuation -> (bool, 'a * 's, 'r, 'f) continuation

| KLoop_in_left : ('a, 's, ('a, 'b) union, 's) kinstr * ('b, 's, 'r, 'f) continuation -> (('a, 'b) union, 's, 'r, 'f) continuation

| KIter : ('a, 'b * 's, 'b, 's) kinstr * 'a list * ('b, 's, 'r, 'f) continuation -> ('b, 's, 'r, 'f) continuation

| KList_enter_body : ('a, 'c * 's, 'b, 'c * 's) kinstr * 'a list * 'b list * int * ('b boxed_list, 'c * 's, 'r, 'f) continuation -> ('c, 's, 'r, 'f) continuation

| KList_exit_body : ('a, 'c * 's, 'b, 'c * 's) kinstr * 'a list * 'b list * int * ('b boxed_list, 'c * 's, 'r, 'f) continuation -> ('b, 'c * 's, 'r, 'f) continuation

| KMap_enter_body : ('a * 'b, 'd * 's, 'c, 'd * 's) kinstr * ('a * 'b) list * ('a, 'c) map * (('a, 'c) map, 'd * 's, 'r, 'f) continuation -> ('d, 's, 'r, 'f) continuation

| KMap_exit_body : ('a * 'b, 'd * 's, 'c, 'd * 's) kinstr * ('a * 'b) list * ('a, 'c) map * 'a * (('a, 'c) map, 'd * 's, 'r, 'f) continuation -> ('c, 'd * 's, 'r, 'f) continuation

| KLog : ('a, 's, 'r, 'f) continuation * logger -> ('a, 's, 'r, 'f) continuation

and ('a, 's, 'b, 'f, 'c, 'u) logging_function = ('a, 's, 'b, 'f) kinstr -> Alpha_context.context -> Alpha_context.Script.location -> ('c, 'u) stack_ty -> ('c * 'u) -> unit

and execution_trace = (Alpha_context.Script.location * Alpha_context.Gas.t * (Alpha_context.Script.expr * string option) list) list

and logger = {

log_interp : a s b f c u. ('a, 's, 'b, 'f, 'c, 'u) logging_function;
log_interp is called at each call of the internal function interp. interp is called when starting the interpretation of a script and subsequently at each Exec instruction.

log_entry : a s b f. ('a, 's, 'b, 'f, 'a, 's) logging_function;
log_entry is called before executing each instruction but after gas for this instruction has been successfully consumed.

log_control : a s b f. ('a, 's, 'b, 'f) continuation -> unit;
log_control is called before the interpretation of the current continuation.

log_exit : a s b f c u. ('a, 's, 'b, 'f, 'c, 'u) logging_function;
log_exit is called after executing each instruction.

get_log : unit -> execution_trace option Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t;
get_log allows to obtain an execution trace, if any was produced.

}

and 'ty ty =

| Unit_t : type_annot option -> unit ty

| Int_t : type_annot option -> Alpha_context.Script_int.z Alpha_context.Script_int.num ty

| Nat_t : type_annot option -> Alpha_context.Script_int.n Alpha_context.Script_int.num ty

| Signature_t : type_annot option -> Alpha_context.signature ty

| String_t : type_annot option -> string ty

| Bytes_t : type_annot option -> bytes ty

| Mutez_t : type_annot option -> Alpha_context.Tez.t ty

| Key_hash_t : type_annot option -> Alpha_context.public_key_hash ty

| Key_t : type_annot option -> Alpha_context.public_key ty

| Timestamp_t : type_annot option -> Alpha_context.Script_timestamp.t ty

| Address_t : type_annot option -> address ty

| Bool_t : type_annot option -> bool ty

| Pair_t : ('a ty * field_annot option * var_annot option) * ('b ty * field_annot option * var_annot option) * type_annot option -> ('a, 'b) pair ty

| Union_t : ('a ty * field_annot option) * ('b ty * field_annot option) * type_annot option -> ('a, 'b) union ty

| Lambda_t : 'arg ty * 'ret ty * type_annot option -> ('arg, 'ret) lambda ty

| Option_t : 'v ty * type_annot option -> 'v option ty

| List_t : 'v ty * type_annot option -> 'v boxed_list ty

| Set_t : 'v comparable_ty * type_annot option -> 'v set ty

| Map_t : 'k comparable_ty * 'v ty * type_annot option -> ('k, 'v) map ty

| Big_map_t : 'k comparable_ty * 'v ty * type_annot option -> ('k, 'v) big_map ty

| Contract_t : 'arg ty * type_annot option -> 'arg typed_contract ty

| Sapling_transaction_t : Alpha_context.Sapling.Memo_size.t * type_annot option -> Alpha_context.Sapling.transaction ty

| Sapling_state_t : Alpha_context.Sapling.Memo_size.t * type_annot option -> Alpha_context.Sapling.state ty

| Operation_t : type_annot option -> operation ty

| Chain_id_t : type_annot option -> Tezos_protocol_environment.Chain_id.t ty

| Never_t : type_annot option -> never ty

| Bls12_381_g1_t : type_annot option -> Tezos_protocol_environment.Bls12_381.G1.t ty

| Bls12_381_g2_t : type_annot option -> Tezos_protocol_environment.Bls12_381.G2.t ty

| Bls12_381_fr_t : type_annot option -> Tezos_protocol_environment.Bls12_381.Fr.t ty

| Ticket_t : 'a comparable_ty * type_annot option -> 'a ticket ty

and ('top_ty, 'resty) stack_ty =

| Item_t : 'ty ty * ('ty2, 'rest) stack_ty * var_annot option -> ('ty, 'ty2 * 'rest) stack_ty

| Bot_t : (empty_cell, empty_cell) stack_ty

and ('key, 'value) big_map = {

id : Alpha_context.Big_map.Id.t option;

diff : ('key, 'value) big_map_overlay;

key_type : 'key comparable_ty;

value_type : 'value ty;

}

and ('a, 's, 'r, 'f) kdescr = {

kloc : Alpha_context.Script.location;

kbef : ('a, 's) stack_ty;

kaft : ('r, 'f) stack_ty;

kinstr : ('a, 's, 'r, 'f) kinstr;

}

and ('a, 's) kinfo = {

iloc : Alpha_context.Script.location;

kstack_ty : ('a, 's) stack_ty;

}

and (_, _, _, _, _, _, _, _) stack_prefix_preservation_witness =

| KPrefix : ('y, 'u) kinfo * ('c, 'v, 'd, 'w, 'x, 's, 'y, 'u) stack_prefix_preservation_witness -> ('c, 'v, 'd, 'w, 'a, 'x * 's, 'a, 'y * 'u) stack_prefix_preservation_witness

| KRest : ('a, 's, 'b, 'u, 'a, 's, 'b, 'u) stack_prefix_preservation_witness

and ('before, 'after) comb_gadt_witness =

| Comb_one : ('a * ('x * 'before), 'a * ('x * 'before)) comb_gadt_witness

| Comb_succ : ('before, 'b * 'after) comb_gadt_witness -> ('a * 'before, ('a * 'b) * 'after) comb_gadt_witness

and ('before, 'after) uncomb_gadt_witness =

| Uncomb_one : ('rest, 'rest) uncomb_gadt_witness

| Uncomb_succ : ('b * 'before, 'after) uncomb_gadt_witness -> (('a * 'b) * 'before, 'a * 'after) uncomb_gadt_witness

and ('before, 'after) comb_get_gadt_witness =

| Comb_get_zero : ('b, 'b) comb_get_gadt_witness

| Comb_get_one : ('a * 'b, 'a) comb_get_gadt_witness

| Comb_get_plus_two : ('before, 'after) comb_get_gadt_witness -> ('a * 'before, 'after) comb_get_gadt_witness

and ('value, 'before, 'after) comb_set_gadt_witness =

| Comb_set_zero : ('value, _, 'value) comb_set_gadt_witness

| Comb_set_one : ('value, 'hd * 'tl, 'value * 'tl) comb_set_gadt_witness

| Comb_set_plus_two : ('value, 'before, 'after) comb_set_gadt_witness -> ('value, 'a * 'before, 'a * 'after) comb_set_gadt_witness

and (_, _) dup_n_gadt_witness =

| Dup_n_zero : ('a * 'rest, 'a) dup_n_gadt_witness

| Dup_n_succ : ('stack, 'b) dup_n_gadt_witness -> ('a * 'stack, 'b) dup_n_gadt_witness

val kinfo_of_kinstr : a s b f. ('a, 's, 'b, 'f) kinstr -> ('a, 's) kinfo

type kinstr_rewritek = {

apply : b u r f. ('b, 'u, 'r, 'f) kinstr -> ('b, 'u, 'r, 'f) kinstr;

}

val kinstr_rewritek : a s r f. ('a, 's, 'r, 'f) kinstr -> kinstr_rewritek -> ('a, 's, 'r, 'f) kinstr