Module Storage_functors.Make_carbonated_data_set_storage

Parameters

module C : Raw_context.T

module I : INDEX

Signature

type t = C.t

type context = t

type elt = I.t

The type of elements.

val mem : context -> elt -> (Raw_context.t * bool) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Tells whether an elt is a member of the set. Consumes Gas_repr.read_bytes_cost Z.zero

val init : context -> elt -> (Raw_context.t * int) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Adds an elt as a member of the set. Consumes Gas_repr.write_bytes_cost <size of the new value>. Returns the the new size.

val remove : context -> elt -> (Raw_context.t * int * bool) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Removes an elt from the set ; does nothing if not a member. Consumes Gas_repr.write_bytes_cost Z.zero. Returns the freed size, and a boolean indicating if a value was already associated to this key.

val fold_keys_unaccounted : context -> init:'acc -> f:(elt -> 'acc -> 'acc Tezos_protocol_environment.Lwt.t) -> 'acc Tezos_protocol_environment.Lwt.t