Module Make.Voting_period_repr

type kind =

| Proposal	(*
protocols can be proposed
*)

| Exploration	(*
a proposal can be voted
*)

| Cooldown	(*
a delay before the second vote of the Promotion period.
*)

| Promotion	(*
activation can be voted
*)

| Adoption	(*
a delay before activation
*)

The voting period kinds are ordered as follows: Proposal -> Testing_vote -> Testing -> Promotion -> Adoption. This order is the one used be the function succ below.

val kind_encoding : kind Tezos_protocol_environment.Data_encoding.t

type voting_period = {

index : Tezos_protocol_environment.Int32.t;

kind : kind;

start_position : Tezos_protocol_environment.Int32.t;

}

A voting period can be of several kinds and is uniquely identified by the counter 'index'. The 'start_position' represents the relative position of the first level of the period with respect to the first level of the Alpha family of protocols.

type t = voting_period

type info = {

voting_period : t;

position : Tezos_protocol_environment.Int32.t;

remaining : Tezos_protocol_environment.Int32.t;

}

Information about a block with respect to the voting period it belongs to: the voting period, the position within the voting period and the number of remaining blocks till the end of the period. The following invariant is satisfied: `position + remaining + 1 = blocks_per_voting_period`

val root : start_position:Tezos_protocol_environment.Int32.t -> t

include Tezos_protocol_environment.Compare.S with type t := voting_period

val (=) : voting_period -> voting_period -> bool

val (<>) : voting_period -> voting_period -> bool

val (<) : voting_period -> voting_period -> bool

val (<=) : voting_period -> voting_period -> bool

val (>=) : voting_period -> voting_period -> bool

val (>) : voting_period -> voting_period -> bool

val compare : voting_period -> voting_period -> int

val equal : voting_period -> voting_period -> bool

val max : voting_period -> voting_period -> voting_period

val min : voting_period -> voting_period -> voting_period

val encoding : t Tezos_protocol_environment.Data_encoding.t

val info_encoding : info Tezos_protocol_environment.Data_encoding.t

val pp : Tezos_protocol_environment.Format.formatter -> t -> unit

val pp_info : Tezos_protocol_environment.Format.formatter -> info -> unit

val pp_kind : Tezos_protocol_environment.Format.formatter -> kind -> unit

val reset : t -> start_position:Tezos_protocol_environment.Int32.t -> t

reset period ~start_position increment the index by one and set the kind to Proposal which is the period kind that start the voting process. start_position is the level at wich this voting_period started.

val succ : t -> start_position:Tezos_protocol_environment.Int32.t -> t

succ period ~start_position increment the index by one and set the kind to its successor. start_position is the level at which this voting_period started.

val position_since : Level_repr.t -> t -> Tezos_protocol_environment.Int32.t

val remaining_blocks : Level_repr.t -> t -> blocks_per_voting_period:Tezos_protocol_environment.Int32.t -> Tezos_protocol_environment.Int32.t